refactor campaign launch package router
